Move image utils into libstorage
This commit creates a new "image.go" file in libstorage so we move image and archive-manipulation functions into it
